Show copyright based on $output->hasCopyright()
Introducing setCopyright/hasCopyright in OutputPage to
make the showing of the copyright message controllable
by pages other than articles.

hasCopyright defaults to isArticle() to preserve the rule
that all article should show the copyright.

This is used by Flow to show the copyright
on various pages that contain user-generated
content.

Bug: T184960
Change-Id: I3a50dbcedc6b119b9262c50cb3a84b0dd230fb3d