resources: Update CLDRPluralRuleParser to v1.3.2-pre
Not yet tagged as release, but updating ahead of that for the
performance improvement that landed with
https://github.com/santhoshtr/CLDRPluralRuleParser/commit/a7782ca.

This is particularly impactful on calls to mw.msg() during the
critical path, due to this library being called by jqueryMsg,
which is used by mw.Message#text, which mw.msg() calls by default.

Particularly seen in the startup code from ULS modules.

Bug: T127328
Change-Id: Ic865e7077d1053c65f7ae1633c60bf52104731b7