Augment precision of updatelist time
MediaWiki updater takes track of each updates that have ever been made.
The primary key in the database is forged based on PHP time() which is
in seconds.

When running install.php and update.php, less than a second might have
elapsed. That result in a duplicate key entry:

  A database query error has occurred.
  Query: INSERT  INTO `updatelog` (ul_key,ul_value)
  VALUES ('updatelist-1.27.5-15396974150', ...

  Function: DatabaseUpdater::setAppliedUpdates
  Error: 1062 Duplicate entry 'updatelist-1.27.5-15396974150'
  for key 'PRIMARY'

Switch to microseconds resolution by forging the key with
microtime(true).

Bug: T207241
Change-Id: Ib457c9f66b6c947f212ad87ca364e799c1c75d90