Step 1 of renaming the language code for Norman from nrm to nrf
The language code nrm is the BCP 47 language code for Narom.
The BCP 47 language code for Norman is nrf.

This change adds support for the language code nrf as fallback language
of the language code nrm. This allows to rename the message files and
the system messages without changing the language code yet.

The next step is to rename the language code in Translatewiki as
described in https://translatewiki.net/wiki/Renaming_language_codes
Also local system messages in MediaWiki:*/nrm can moved to
MediaWiki:*/nrf.

The final step is the change Ib8d3c344ebe082e4d3e5ec12b3815c5845f4797e
to complete the rename of the language codes.

Bug: T25216
Bug: T165648
Change-Id: If220cc7f094672389c8e6cc331f753df067503f2