Fix comment handling on image upload or deletion
Before Iab5f5215, the call to CommentStore::insertWithTempTable() also
happened to populate image_comment_temp for the later call to
insertSelect() when moving rows from the image table to oldimage or
filearchive. There was nothing in the image table itself that needed
updating.

In that change those calls were changed to CommentStore::insert(), but
it was missed that in that case we do have to update the image table
itself.

Bug: T207419
Change-Id: I26c417c9ab8a9160a7c7ec548ffdfabf17f01980