Rename WebInstallerOutput::addWikiTextInterface
Follow up to I89f3398cffa771afcd5a33cfd11eb8510af3e7f7: this patch
was intended to mirror a new method added to OutputPage in
Ia58910164baaca608cea3b24333b7d13ed773339 but the bikeshed was painted
a slightly different color before the OutputPage patch was merged, and
somehow this patch wasn't changed to match.  The `addWikiTextInterface`
name was never included in a released MW version, so we don't need to
deprecate it, just fix it.

Change-Id: I636cb0ea0f9631c4c5e2bc118f479aec76d2b231