RCFilters: Fix highlighted rows overlapping legend
While the new rows are being faded in, they each have an opacity not
equal to 1. Setting the opacity on something creates a new stacking
context, and because the rows are below the legend in the DOM, this made
them appear on top of the legend.

Fix this by setting a z-index on the legend so it stays on top.

Bug: T186841
Change-Id: I2eee201d1e0aed94974a85d6d46b86fd4a4db796