Changelist "tags" item is wrapped in brackets via CSS
In our quest to separate content from presentation,
the parentheticals are moved into CSS. These support translation via
the lessMessages ResourceLoader property.

There is obviously a risk with this change of double brackets appearing.
We might want to consider a new message to avoid this (at the cost of
some translations being outdated). However, I believe it's better to
have the double paranthetical wrap then to have English strings show up
in the wrong places. We can update other translations if
necessary.

Bug: T205581
Change-Id: I9c3fb069fc93800f39a922feab56489f1614785c