Make the context page the edited page in EditPages
This forces the WikiPage returned by $this->getContext()->getPage()
to be the same as $this->page.

This seems to be redundant at the moment, but it seems prudent to
make sure this does not break. Having two WikiPage instances may
cause subtle problems, such as redundant parsing of content during
edits.

Bug: T205369
Change-Id: I8dd3235ec395fa8772ad810f8b09a098a940ab0c