config: Change ConfigException to extend LogicException
This exception is thrown when someone tries to fetch a
configuration entry with an invalid name, or override an already
set entry, both of which are errors in the logic and cannot be
meaningfully handled, so it does not make sense for this exception
to trigger IDE warnings.

(The one exception is the etcd load error, which probably shouldn't
be a ConfigException in the first place. But even that is not
something callers could meaningfully handle.)

Change-Id: I26b5b32f3348965f0cbf607cbb7cb5aaa509ed51