parser: Remove aliases for __EXPECTUNUSEDCATEGORY__
Follows-up 4985ce513469fa82a132139a41c2854504475e4a.

This is a brand new magic word, with no history or concerns for
backward-compatibility. Introducing it with four possible ways to
to use it, I think leads to less ideal experience for end-users.

I recognise that this is an attempt to make it easier for users
to get this magic word correctly working, even when a user guesses
the name incorrectly. This is a valid problem but, I do not think
adding more syntax is the right answer to that problem.

Introducing four ways to do the same thing I believe would make it
more difficult to learn the syntax, because they also have to
remember the other variations when encountering it from other users.

I also believe that the value we gain from a user getting it right
via an alias of this keyword is, shadowed by the loss in value
from a user when they get it wrong for all the other keywords.
This is because they'd likely expect the same four varations
(MULTIWORD, MULTI_WORD, MULTIPLE_WORD, MULTIPLEWORD) to apply
there as well. It creates expectations that we are not matching.

I think it is worth investigating this problem, and having four
extra aliases in the syntax could be one of the proposals to solve
that problem. However, we should discuss that on Phabricator and,
if we agree to it, apply it universally, not just to one keyword.

Bug: T96041
Change-Id: I5621cee70e6f791a161f21ebda8c5a7243e5947e