Allow users to block the user that blocked them.
This is to make it so that blocking all other admins is not
a succesful attack plan, as the blocked admins can block the
blocker, and then it ends in a stalemate with everyone blocked.

This also allows users with unblock-self right to adjust their
own blocks. The code already existed for this but was broken.

Credit for this idea goes to Tgr.

Bug: T150826
Change-Id: I0418279fdb2a59f8f1d7eeb8931d874123d03e4f