Fix RevisionDeleteUser rev_actor query for MySQL
MySQL uses an extremely bad plan for the update with subquery here.
For the time being let's split out the subquery.

This'll break if the user has too many revision rows, but on the other
hand all the existing queries here will probably break in the same way
so let's leave that for later.

Bug: T210628
Change-Id: Ia3eb41b43b96c506349279e19743c41530969636