export: Fix return value of write() function in DumpOutput.php
Per http://php.net/manual/en/function.print.php, print always returns
1 (integer value of one) but write() doesn't actually return this value
to it's caller. So "@return bool" in this case doesn't make sense as
one will think it's returning a bool type. `write()` only takes a string
and prints it, and if we really want to return it's value (in this case),
it will be "@return int" not data of type bool.

Change-Id: I45b4a157cde2f768d77dda76d6ae0caa47e28f20