Special:AllMessages: Remove 'id' attribute from message rows
This was introduced in r13629 (2006; 17633ab92a4) to allow
easy linking to individual results, which predates the pagination
feature that was added in r53286 (2009; c90b9b93ecb).

This was a hidden feature not exposed througuh the user-interface
in any way, and not documented. Removing it should be safe.

Any recently crafted permalinks fallback to loading the page containing
the message key without jumping to it.

Any older permalinks have already been broken for 10 years because
pagination would have limited the results returned, likely not
included the anchored ID (dead-end fragment).

Change-Id: I53bd56e0d5a82fba65388e1c2da3c8349a138261